peer's port from int to uint16_t cuz it's correct

+++ src/peer.h
@@ -2,19 +2,20 @@
 #define PEER_H_
 
 #include <string>
+#include <cstdint>
 
 class Peer {
 private:
     std::string peer_id, ip;
-    int port;
+    uint16_t port;
 public:
-    Peer(std::string peer_id, std::string ip, int port)
+    Peer(std::string peer_id, std::string ip, uint16_t port)
         : peer_id(peer_id),
           ip(ip),
           port(port)
         {};
 
-    Peer(std::string ip, int port)
+    Peer(std::string ip, uint16_t port)
         : peer_id(""),
         ip(ip),
         port(port)
